+++ README.md
@@ -1,5 +1,7 @@
 # Virus Model Simulator
 
+# Introduction
+
 A simulator created by Bruce Yong to simulate spread of 2019-nCov 
 and tell everyone on BiliBili why it's safer to stay home while the coronavirus breaks out.  
 
@@ -9,3 +11,4 @@ The code has been modified, if you want to see the original code. Please go to h
 | --: | :-- |
 | Source Repo | [KikiLetGo/VirusBroadcast](https://github.com/KikiLetGo/VirusBroadcast) |
 | Written in | Java |
+| Original Developer | Bruce Yong |
