+++ src/virussim/Hospital.java
@@ -1,3 +1,5 @@
+package virussim;
+
 import java.util.ArrayList;
 import java.util.List;
