+++ src/virussim/Main.java
@@ -1,3 +1,5 @@
+package virussim;
+
 import javax.swing.*;
 
 import java.awt.Color;
