+++ src/virussim/Person.java
@@ -1,3 +1,5 @@
+package virussim;
+
 import java.util.List;
 import java.util.Random;
